+++ kukumo-plugins/kukumo-html-report/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning][2].
 ### Added
 
 - Result for each feature and scenario
-- Error display by type
+- Categorize test case errors according the step contributor
 
 ### Fixed